% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crayonmaps.R
\name{colorKey}
\alias{colorKey}
\title{Color key}
\usage{
colorKey(values, pal = colorRampPalette(c("red", "gray", "blue"))(24),
  scale = TRUE)
}
\arguments{
\item{values}{vector of numeric values to which colors will be mapped}

\item{pal}{color palette}

\item{scale}{logical, should values be scaled? Defaults to TRUE}
}
\value{
a vector of colors, mapped to the values and in the same order as the values vector
}
\description{
Color key
}
\author{
Giuseppe D'Agostino
}
