% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crayonmaps.R
\name{rotateccw}
\alias{rotateccw}
\title{Rotate a matrix counter-clockwise}
\usage{
rotateccw(x)
}
\arguments{
\item{x}{matrix to be rotated}
}
\value{
a matrix rotated counter-clockwise
}
\description{
Rotate a matrix counter-clockwise
}
\author{
Matthew Lundburg in https://stackoverflow.com/questions/16496210/rotate-a-matrix-in-r
}
