% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crayonmaps.R
\name{textDendro}
\alias{textDendro}
\title{Text-based heatmap}
\usage{
textDendro(clust, xfac = 2, ylim = 20, horiz = FALSE,
  dcol = "white")
}
\arguments{
\item{clust}{hierarchical clustering object return by hclust()}

\item{xfac}{numeric, sets the spacing used to rescale the x coordinates. Default is 2}

\item{ylim}{numeric, sets the maximum amount of y (dendrogram height) used to rescale. Default is 20}

\item{horiz}{logical, is the dendrogram oriented as left to right? Default is FALSE}

\item{dcol}{character, R color name used to draw the dendrogram}
}
\value{
a text-based dendrogram
}
\description{
Text-based heatmap
}
\author{
Alan O'Callaghan and Giuseppe D'Agostino
}
