% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crayonmaps.R
\name{textHeatmap}
\alias{textHeatmap}
\title{Text-based heatmap,}
\usage{
textHeatmap(dat, pal = colorRampPalette(c("red", "gray", "blue"))(24),
  cluster_cols = TRUE, cluster_rows = TRUE,
  dist_method = "euclidean", clustering_method = "complete",
  main = "Heatmap", key = "Key", show_col_legend = TRUE, lpad = "")
}
\arguments{
\item{dat}{matrix with numeric values, rownames and colnames}

\item{cluster_cols}{logical, should columns be clustered by hierarchical clustering? default is TRUE}

\item{cluster_rows}{logical, should rows be clustered by hierarchical clustering? default is TRUE}

\item{dist_method}{character, distance method used by dist}

\item{clustering_method}{character, clustering method used by hclust}

\item{main}{character, title of the heatmap}

\item{key}{character, title of the color legend}

\item{show_col_legend}{logical, should the column legend be shown at the end of the plot? default is TRUE}

\item{cols}{color palette}
}
\value{
a text-based heatmap directly in the terminal output using ANSI background styles. Useful when your X11 forwarding is broken or for quick exploratory analysis of small datasets. Setup is largely inspired by pheatmap.
}
\description{
Text-based heatmap,
}
